Upgrade tooling / Fix recovery issue / Fix testing env
This repository's been suffering a bit of bit-rot, and when I found a
bug in the migrations system I needed to fix, I took some time to do a
few other things.

The core issue is that if you have a series of migrations 0-3, and 3 is
bad and aborts the migration, the current code would have you start back
at 0 the next run, which is wrong since all the way up to 2 have
completed. An ugly oversight that I hope never returns. Tests were added
for this bug.

Other things: upgrade all go tooling, upgrade to latest (future proofed)
postgres packaging.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>